requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coloredlogs==15.0.1
frozenlist==1.3.3
gevent==23.7.0
greenlet==2.0.2
grequests==0.7.0
humanfriendly==10.0
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
psycopg2==2.9.6
pycparser==2.21
pyreadline3==3.4.1
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
SQLAlchemy==2.0.18
typing_extensions==4.7.1
urllib3==2.0.3
yarl==1.9.2
zope.event==5.0
zope.interface==6.0
